This is my best code to solve day 19 part 2
It still sucks since it depends on the sequence in which I apply the regular
expressions. Often it gets stuck on a sequence it can't reduce any further.
When that happens I shuffle the regexes and try again. Eventually it finds a
solution that converges to e. Sometimes it finds a solution that's fewer steps
than what AoC accepted as the "right" answer!